Move methods and tests from board to evaluator
