+++ lib/board.rb
@@ -1,26 +1,19 @@
+require_relative 'evaluator'
+
 class Board
   attr_reader :cells
 
   def initialize(board_hash, board_dimension)
     @cells = board_hash
     @board_dimension = board_dimension
+    @evaluator = Evaluator.new
   end
 
   def valid_placement?(coordinates, ship)
-    coordinates_match_ship_length?(coordinates, ship) &&
-    no_duplicate_coordinates?(coordinates) &&
+    @evaluator.coordinates_match_ship_length?(coordinates, ship) &&
+    @evaluator.no_duplicate_coordinates?(coordinates) &&
     coordinates_empty?(coordinates) &&
-    is_consecutive?(coordinates, ship)
-  end
-
-  def is_consecutive?(coordinates, ship)
-    if is_horizontal?(coordinates)
-      return ((user_coordinate_numbers(coordinates).last - user_coordinate_numbers(coordinates).first) + 1 == ship.length)
-    elsif is_vertical?(coordinates)
-      return ((user_coordinate_letters(coordinates).last.ord - user_coordinate_letters(coordinates).first.ord) + 1 == ship.length)
-    else
-      return false
-    end
+    @evaluator.is_consecutive?(coordinates, ship)
   end
 
   def valid_coordinate?(coordinate)
@@ -31,40 +24,6 @@ class Board
     coordinates.all? do |coordinate|
       @cells[coordinate.to_sym].empty?
     end
-  end
-
-  def coordinates_match_ship_length?(coordinates, ship)
-    coordinates.count == ship.length
-  end
-
-  def no_duplicate_coordinates?(coordinates)
-    coordinates.uniq.count == coordinates.length
-  end
-
-  def split_user_coordinates(coordinates)
-    coordinates.map do |coordinate|
-      coordinate.split("")
-    end
-  end
-
-  def user_coordinate_numbers(coordinates)
-    split_user_coordinates(coordinates).map do |sub_arr|
-      sub_arr[1].to_i
-    end.sort
-  end
-
-  def user_coordinate_letters(coordinates)
-    split_user_coordinates(coordinates).map do |sub_arr|
-      sub_arr[0]
-    end.sort
-  end
-
-  def is_horizontal?(coordinates)
-    user_coordinate_letters(coordinates).uniq.count == 1
-  end
-
-  def is_vertical?(coordinates)
-    user_coordinate_numbers(coordinates).uniq.count == 1
   end
 
   def place(coordinates, ship)
